backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.3.5
langchain-community>=0.0.10
pypdf>=3.17.0
PyMuPDF>=1.24.0
pillow>=10.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
